feat(blog-section): refactor blog-entries in reusable components
