+++ src/components/UI/blog-entries/utils.js
@@ -20,6 +20,4 @@ const generateDistribution = (cardsLength) => {
   });
 };
 
-export {
-  generateDistribution,
-}
+export { generateDistribution };
